[ broken ] Making {-# OPTIONS --safe #-} valid
This is meant to address #1678 and it does work when used from
the command line.

However because the emacs mode seems to set --allow-unsolved-metas
by default in order for the interaction to work, it completely
breaks the emacs interface: as soon as one tries to set --safe
in a pragma, Agda complains about the incompatibility of the two
options.

Could this be a good use case to revive --interactive? It would
basically allow unsolved metas but only because the user is
interacting via a text editor and not because unsolved metas are
deemed intrinsically acceptable.